move timeline example into its own component
