+++ app/page.module.css
@@ -3,17 +3,6 @@
 	max-width: 100%;
 }
 
-.darkmode {
-	background-color: rgb(22, 30, 39);
-}
-
-.embedexample {
-	height: 800px;
-	overflow-y: scroll;
-	padding: 10px;
-	width: 500px;
-}
-
 .embedcode {
 	max-width: 100%;
 	width: 100%;
