+++ components/TimelineExample.module.css
@@ -1,0 +1,10 @@
+.darkmode {
+	background-color: rgb(22, 30, 39);
+}
+
+.embedexample {
+	height: 800px;
+	overflow-y: scroll;
+	padding: 10px;
+	width: 500px;
+}
