--- util/generatecustomcss.ts
@@ -1,6 +1,0 @@
-import { ColorList } from '@/types/data';
-
-const generateCustomCSS = (styles: ColorList) =>
-	`#embedbsky-com-timeline-embed .postcontainer {background:${styles.background};}#embedbsky-com-timeline-embed .postcontainer .repostheader {color:${styles.repostHeader};}#embedbsky-com-timeline-embed .postcontainer .repostheader a {color:${styles.repostHeader};}#embedbsky-com-timeline-embed .postcontainer .postbox {border-bottom: 1px solid ${styles.border};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text {color: var(--v-color-text);}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text strong a:link,#embedbsky-com-timeline-embed .postcontainer .postbox .col.text strong a:visited,#embedbsky-com-timeline-embed .postcontainer .postbox .col.text strong a:active,#embedbsky-com-timeline-embed .postcontainer .postbox .col.text strong a:hover {color:${styles.linkName};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text strong a:hover {color:${styles.linkNameHover};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .handle {color:${styles.linkHandle};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .handle a:link,#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .handle a:visited,#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .handle a:active,#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .handle a:hover {color:${styles.linkHandle};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .handle a:hover {color:${styles.linkHandle};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .timeago a:link,#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .timeago a:visited,#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .timeago a:active,#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .timeago a:hover {color:${styles.linkTimestamp};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .timeago a:hover {color:${styles.linkTimestampHover};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .textcopy a:link,#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .textcopy a:visited,#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .textcopy a:active {color:${styles.link};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .textcopy a:hover {color:${styles.linkHover};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .linkcard {border: 1px solid ${styles.border};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .linkcard a {color:${styles.linkLinkCard};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .linkcard .site {color:${styles.linkHandle};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .icons .num {color:${styles.counts};}#embedbsky-com-timeline-embed .postcontainer .postbox .col.text .quotebox {border: 1px solid ${styles.border};}`;
-
-export default generateCustomCSS;
